feat(grunt-conventional-changelog): introduces grunt-conventional-changelog plugin
this plugin let us auto generate changelogs for this project.
all you have to do is to provide proper commit messages that
follow a given convention (like this one).

the angular project itself uses conventional changelogs to
automate this process.
to generate a changelog simply grunt [4mRunning "changelog" task[24m and
specify a release tag from  which the changelog
should be generated.