Add `manageScope` selector function
Allows any old observable sequence to be used in angular context. This
means that observables using this transform function are automatically
observing on the ScopeScheduler and are disposed upon
`$scope.$on("destroy")`.

This has the following advantages:

+ All code in `.subscribe` is happening in the right time of the digest
  cycle, i.e no flickering or missed updates
+ When the scope is destroyed, the observable is disposed off. This
  avoids subtle memory leaks where members of the scope managed by
  `ScopeScheduler` are still referenced past the scope's destruction.

This function is to be used in conjunction with with
`Rx.Observable.prototype.let`, consider:

```
Rx.Observable.interval(100)
  .let(Rx.manageScope($scope))
  .subscribe(...)
;
```